Several tweaks to unitful handling
This proposes several tweaks on top of #35:

1. fixes a couple places where `ustrip` was used without `uconvert`,
which could cause issues for unitless quantities with powers-of-ten,
e.g. `1kHz*1s`
2. `frames` are always zero-indexed. The idea here is to make them more
consistent with other unitful quantities, as they're always used as an
offset or duration. Also often when doing DSP it's convenient to operate
with zero-indexed buffers, so this gives the user an easy way do do
that.
3. makes `inframes`, `inHz`, and `inseconds` more uniform, in that they
all convert to a real-valued unitless value and don't do any rounding.
This also makes the function of `inframes` (which is responsible for
unit conversion) more distinct from `toindex` (which is responsible for
generating a valid native index value). When reviewing #35 they seemed a
little intermingled.

some of the unit conversion for stream handling is still a little messy,
but I think that's going to have to be reworked a bit to fix #18 and #7
anyways.

@haberdashPI do these seem like reasonable tweaks to you?