RFC: proposed reimplementation to handle more graceful signal reformatting.
This is a draft proposal to really mix things up. All the existing
tests pass locally, but I have yet to write any new tests.

It defines the basis of a new interface which could be used to
build the functionality present in the `Sounds` package.

The formatting is largely handled through a new trait called SignalFormat.
This is the basis of a function, `promote_signals` which can be used to ensure
all signals passed are an appropriate format.