Fix downgrade of packages
I think this is correct syntax but it's not tested.

I was adding LibSndFile but this package seems to be the cause (or Compat). I think their reason for version 3 is just to simplify code (issue 651), dropping support for Julia 0.7, and you would keep it if you allow Compat 2.x.